+++ .gitignore
@@ -1,3 +1,5 @@
+**/.env
+
 **/target
 **/dist
