--- backend/.env
@@ -1,2 +1,0 @@
-IPV4='127.0.0.1'
-PORT='3456'
