+++ backend/config_model.yaml
@@ -1,4 +1,4 @@
-cpu: true
+cpu: false
 use_flash_attn: false
 model_id: "lmz/candle-mistral"
 revision: "main"
